refactor : profile type 루트로 이동
