+++ app/join/funnels/step5/components/PersonalitySection.client.tsx
@@ -1,7 +1,7 @@
 import PersonalityItemServer from './PersonalityItem.client';
 import { personalityList } from '@/constants/personalityList';
 
-import type { PersonalityType } from '@/types/profile';
+import type { PersonalityType } from '@/types';
 
 export default function PersonalitySection() {
   return (
