--- types/profile.ts
@@ -1,5 +1,0 @@
-export type PersonalityType = {
-  id: number;
-  keyword: string;
-  keywordInEnglish: string;
-};
